Implement view changes
PLCR18SW-7

+++ Packlink/Services/BusinessLogic/ShopShippingMethodService.php
@@ -109,7 +109,9 @@ class ShopShippingMethodService implements BaseService
             }
         }
 
-        $this->getBaseRepository()->delete($map);
+        if ($map) {
+            $this->getBaseRepository()->delete($map);
+        }
 
         return true;
     }
